crypto_deploy_complete/requirements-docker.txt
nltk>=3.9.1
numpy>=2.2.4
openai>=1.75.0
pandas>=2.2.3
pandas-ta>=0.3.14b0
plotly>=6.0.1
psycopg2-binary>=2.9.10
requests>=2.32.3
streamlit>=1.44.1
trafilatura>=2.0.0